Fix: [Referenda Tracks] Resolve representation issues that are breaking PJS apps (#7671)
This Pull Request fixes the following issues present with PJS derived
from #2072

- Represents `Tracks` constant as a `Vec<(TrackId, TrackInfo)>` instead
of `Vec<Track>`, so tracks iteration in PJS apps _Governance_ hook
remains as how it was before, and

![Screenshot 2025-02-22 at 12 31
28 pm](https://github.com/user-attachments/assets/a075b0b2-8984-43cc-9e72-99996de0ae1b)

- Encapsulates `[u8; N]` from `name` field in a `StringLike` structure
that represents its `TypeInfo` as a `&str` and encodes as such. This
fixes errors present in PJS apps that treat `name` as a javascript
`string`.

![Screenshot 2025-02-22 at 12 31
18 pm](https://github.com/user-attachments/assets/361a4f10-47b7-496d-9591-bc37afcf5ee1)

---------

Co-authored-by: Guillaume Thiolliere <gui.thiolliere@gmail.com>
Co-authored-by: Oliver Tale-Yazdi <oliver.tale-yazdi@parity.io>